Refactor: move CachePolicy to akka.io.dns.iternal, and wrap ADT in object
