+++ akka-actor-tests/src/test/scala/akka/io/dns/internal/AsyncDnsResolverSpec.scala
@@ -8,16 +8,16 @@ import java.net.{ Inet6Address, InetAddress }
 
 import akka.actor.Status.Failure
 import akka.actor.{ ActorRef, ExtendedActorSystem, Props }
-import akka.io.Ttl
+import akka.io.dns.DnsProtocol._
+import akka.io.dns.internal.AsyncDnsResolver.ResolveFailedException
+import akka.io.dns.internal.CachePolicy.Ttl
+import akka.io.dns.internal.DnsClient.{ Answer, Question4, Question6, SrvQuestion }
 import akka.io.dns.{ AAAARecord, ARecord, DnsSettings, SRVRecord }
 import akka.testkit.{ AkkaSpec, ImplicitSender, TestProbe }
 import com.typesafe.config.ConfigFactory
-import akka.io.dns.DnsProtocol._
-import akka.io.dns.internal.AsyncDnsResolver.ResolveFailedException
-import akka.io.dns.internal.DnsClient.{ Answer, Question4, Question6, SrvQuestion }
 
-import scala.concurrent.duration._
 import scala.collection.{ immutable ⇒ im }
+import scala.concurrent.duration._
 
 class AsyncDnsResolverSpec extends AkkaSpec(
   """
